changelog/ implementation/requirements.txt
# CloudBrush 汉字图片搜索系统 - Python 依赖
# 
# 安装: pip install -r requirements.txt

# 核心依赖
mitmproxy==10.1.6           # HTTP/HTTPS 抓包工具
boto3==1.34.12              # AWS SDK (用于 R2/S3 上传)
requests==2.31.0            # HTTP 请求库

# 进度条和可视化
tqdm==4.66.1                # 进度条显示

# 数据处理
Pillow==10.1.0              # 图片处理（可选，用于生成字体图片）

# 开发工具（可选）
pytest==7.4.3               # 测试框架
black==23.12.0              # 代码格式化
flake8==6.1.0               # 代码检查

# ============================================================================
# 最小安装（仅核心功能）
# ============================================================================
# pip install mitmproxy boto3 tqdm

# ============================================================================
# 版本说明
# ============================================================================
# 
# mitmproxy: 
#   - 用于抓包采集 CloudBrush 图片
#   - 需要 Python 3.9+
#   - 安装后使用: mitmproxy, mitmweb, mitmdump
# 
# boto3:
#   - AWS SDK，用于上传到 S3 兼容存储（R2/S3）
#   - 支持 Cloudflare R2 和 AWS S3
# 
# requests:
#   - 用于 HTTP 请求和测试
# 
# tqdm:
#   - 显示上传进度条
#   - 提升用户体验
# 
# Pillow (可选):
#   - 如果使用开源字体生成图片
#   - 需要额外字体文件
# 
# ============================================================================
# 安装说明
# ============================================================================
# 
# 1. 创建虚拟环境（推荐）
#    python3 -m venv venv
#    source venv/bin/activate  # Linux/Mac
#    venv\Scripts\activate     # Windows
# 
# 2. 安装依赖
#    pip install -r requirements.txt
# 
# 3. 验证安装
#    mitmproxy --version
#    python3 -c "import boto3; print(boto3.__version__)"
# 
# ============================================================================
# 故障排除
# ============================================================================
# 
# Q: mitmproxy 安装失败？
# A: 需要 Python 3.9+，升级 Python 版本
# 
# Q: 依赖冲突？
# A: 使用虚拟环境隔离依赖
# 
# Q: 网络问题？
# A: 使用国内镜像源
#    pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
# 
# ============================================================================
